Halve generation RAM: complex64 + window-aware fldigi post-proc cap
OOM root cause at window_len=12000:
fldigi.run() hardcoded postproc_pool = ThreadPoolExecutor(max_workers=3)
which was sized when each in-flight mode held ~1.5 GB at the old 2048
window length.  At 12000 samples each in-flight mode allocates
~9 GB (raw_windows + apply_impairments output, both complex128) and
3 concurrent post-procs blew past 86 GB RAM, getting OOM-killed.

Changes:

1. fldigi.py: postproc_pool concurrency now scales by per-mode RAM
   estimate.  Budget 6 GB per concurrent post-proc, computed from the
   max boosted samples_per_class × window_len × 8B (complex64) × 2.
   At window_len=2048 → 3 workers (matches old behavior).
   At window_len=12000 → 1 worker (was 3, cause of the OOM).

2. extract_windows() takes a `dtype` kwarg (default complex128 for
   backward compat).  fldigi/analog/digivoice now pass complex64,
   halving raw_windows footprint from 16B/sample to 8B/sample.

3. apply_impairments() calls in fldigi/analog/digivoice now pass
   dtype=complex64 explicitly so the output array is also halved.

4. base.py BaseGenerator dtype fallback flipped from complex128 to
   complex64 when no domain matches (e.g. fs=12000 + window_len=12000).
   Domain hits still use the declared dtype, so narrowband 2048 stays
   complex128 — no change for existing 2048 datasets.  base.py now
   also passes self._dtype to extract_windows so the wsjtx/sstv/packet
   path matches the rest.

5. cli.py rewindow path: when window_len differs from the resolved
   domain, use complex64 instead of the domain's declared dtype.

Combined effect at window_len=12000, n_samples=22500:
  before: raw=4.32 GB + samples=4.32 GB = 8.64 GB per mode × 3 = 26 GB
  after:  raw=2.16 GB + samples=2.16 GB = 4.32 GB per mode × 1 =  4 GB

Training pipeline already converts to float32, so the precision drop
is invisible downstream.

+++ src/rf_datagen/generators/analog.py
@@ -171,7 +171,8 @@ class AnalogGenerator(BaseGenerator):
                 raw_windows = extract_windows(
                     combined_iq, window_len=self.window_len,
                     stride=stride, power_threshold=power_threshold,
-                    max_windows=n_samples)
+                    max_windows=n_samples,
+                    dtype=np.complex64)
                 del combined_iq
                 if len(raw_windows) == 0:
                     log.warning("%15s: FAILED (no valid windows)", mode_name)
@@ -181,7 +182,8 @@ class AnalogGenerator(BaseGenerator):
 
                 samples, meta = apply_impairments(
                     raw_windows, n_samples, fs=self.fs,
-                    window_len=self.window_len, return_metadata=True)
+                    window_len=self.window_len, return_metadata=True,
+                    dtype=np.complex64)
 
                 atomic_save_npy(npy_path, samples)
                 snrs = meta.get("snrs", [])
